pyproject.toml
[tool.poetry]
name = "bigami-sheet-api"
version = "0.1.0"
description = ""
authors = ["ac2393921 <ac2393921@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.1"
strawberry-graphql = "^0.198.0"
uvicorn = "^0.23.2"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pydantic = "1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
